requirements.txt
langchain
langchain-community
langchain-groq
chromadb
sentence-transformers
duckduckgo-search
python-dotenv
streamlit
